moved the pbr render pass outside of scene.cpp because scene.cpp should just be concerned w object hierarchy...render passes can rely on the scene iterator instead
