+++ resources/models/rigged_figure/RiggedFigure.gltf
@@ -14,8 +14,8 @@
     "nodes": [
         {
             "children": [
-                21,
-                1
+                1,
+                21
             ],
             "matrix": [
                 1.0,
