+++ src/gfx_api/gfx.h
@@ -279,11 +279,6 @@ void unbind_framebuffer();
 
 void get_gfx_error();
 
-struct obj_sort_info_t {
-  int obj_id = -1;
-  vec3 pos; 
-};
-
 enum RENDER_MODE {
 	NORMAL,
 	WIREFRAME
